correct mGBA source package
I rebuilt mGBA because the upgrade from Debian 9 to 10 had caused it
to link against a no-longer-present major version of libavformat.
I tried this:

$ sudo apt build-dep mgba-qt

I got a warning:

Picking 'mgba' as source package instead of 'mgba-qt'

I'm guessing it prefers the name of the source package over the name
of the binary package produced from it.  This change might shortcut
the warning, or at worst it might fail to install the prereqs for
the Qt front end.  We'll see once Debian 11 comes out.